refactor: remane file with interfaces
