+++ rss-puzzle/src/types/interfaces.ts
